+++ tests/testthat/test-spglm.R
@@ -7,9 +7,13 @@ test_that("generics work spglm point data", {
   spmod1 <- spglm(abs(y) ~ x, "Gamma", exdata, spcov_type = "exponential", xcoord = "xcoord", ycoord = "ycoord", estmethod = "reml")
   spmod2 <- spglm(abs(y) ~ x, Gamma, exdata, spcov_type = "none", xcoord = xcoord, ycoord = ycoord, estmethod = "reml")
 
-  # AIC
+  # AIC, AICc, BIC
   expect_vector(AIC(spmod1))
   expect_s3_class(AIC(spmod1, spmod2), "data.frame") # turn reml fixed effects warning off
+  expect_vector(AICc(spmod1))
+  expect_s3_class(AICc(spmod1, spmod2), "data.frame")
+  expect_vector(BIC(spmod1))
+  expect_s3_class(BIC(spmod1, spmod2), "data.frame")
 
   # anova
   expect_s3_class(anova(spmod1), "data.frame")
@@ -85,6 +89,7 @@ test_that("generics work spglm point data", {
 
   # loocv
   expect_vector(loocv(spmod1))
+  expect_type(loocv(spmod1, cv_predict = TRUE, type = "response"), "list")
   expect_type(loocv(spmod1, cv_predict = TRUE, se.fit = TRUE, local = TRUE), "list")
 
   # model.frame
@@ -158,9 +163,13 @@ test_that("generics work spglm point data with missing", {
   spmod1 <- spglm(abs(y) ~ x, family = "Gamma", data = exdata_M, spcov_type = "exponential", xcoord = "xcoord", ycoord = "ycoord", estmethod = "reml")
   spmod2 <- spglm(abs(y) ~ x, family = Gamma, data = exdata_M, spcov_type = "none", xcoord = xcoord, ycoord = ycoord, estmethod = "reml")
 
-  # AIC
+  # AIC, AICc, BIC
   expect_vector(AIC(spmod1))
   expect_s3_class(AIC(spmod1, spmod2), "data.frame") # turn reml fixed effects warning off
+  expect_vector(AICc(spmod1))
+  expect_s3_class(AICc(spmod1, spmod2), "data.frame")
+  expect_vector(BIC(spmod1))
+  expect_s3_class(BIC(spmod1, spmod2), "data.frame")
 
   # anova
   expect_s3_class(anova(spmod1), "data.frame")
@@ -236,6 +245,7 @@ test_that("generics work spglm point data with missing", {
 
   # loocv
   expect_vector(loocv(spmod1))
+  expect_type(loocv(spmod1, cv_predict = TRUE, type = "response"), "list")
   expect_type(loocv(spmod1, cv_predict = TRUE, se.fit = TRUE, local = TRUE), "list")
 
   # model.frame
@@ -386,6 +396,7 @@ test_that("generics work spglm polygon data with missing", {
 
   # loocv
   expect_vector(loocv(spmod1))
+  expect_type(loocv(spmod1, cv_predict = TRUE, type = "response"), "list")
   expect_type(loocv(spmod1, cv_predict = TRUE, se.fit = TRUE, local = TRUE), "list")
 
   # model.frame